+++ openeditor/lib.py
